requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
psycopg2-binary
Pillow
drf-yasg
coreapi
django-rest-swagger
whitenoise